# perf: Cache transaction on propose and validate to replay on commit